+++ CHANGELOG.md
@@ -30,6 +30,7 @@ Release Versions:
 - feat: update to jazzy and CL v9.0.0 (#116)
 - refactor: move all exceptions into one place (#117)
 - feat(controllers): add flag for collision features to robot controller interface (#114)
+- feat: try catch lifecycle transitions in components and controllers (#120)
 
 ## 4.2.2